more random

+++ src/main.c
@@ -135,7 +135,7 @@ void predict(period *data, period *prediction) {
         for (int j = i + 1; j < 6; ++j) {
 		while (prediction->balls[i] == prediction->balls[j]) {
 			int direction = rand() >= 0.5 ? -1 : 1;
-			int times = rand() * 3;
+			double times = rand() * 3;
 			prediction->balls[j] = (int)(prediction->balls[j] + direction * times * stddev[j] - direction * 1) % 33;
 			if (prediction->balls[j] < 0) prediction->balls[j] += 33;
 			prediction->balls[j] += 1;
